refactor: remove unused OCR and chunking modules, and add new PDF handling and evaluation scripts
