+++ evaluation/improve_generated_qa.py
@@ -12,7 +12,7 @@ from glob import glob
 
 from tqdm import tqdm
 
-from ..utilities.llm_models import get_llm_model_chat
+from src.utilities.llm_models import get_llm_model_chat
 from .prompts import IMPROVE_QA, IMPROVE_QA_CONTENT
