+++ ocr/README.md
