+++ ocr/config/ocr_config.py
